Fixed the all outs = 0 problem; test file basis
The otsu_sigma_b function would always return an array of zeros. This happened because _apply_scalar_per_pixel casts its output dtype to that of its input (i.e. uint8) by default, and the outputs were too large to be represented by uint8 (so the function returned 0).

(I also laid the groundwork for my test file. I need to add the manually computed solution, then it should be good to go).

+++ skimage/filters/rank/generic.py
@@ -1440,7 +1440,8 @@ def otsu_sigma_b(image, selem, out=None, mask=None,
     if np_image.ndim == 2:
         return _apply_scalar_per_pixel(generic_cy._otsu_sigma_b, image, selem,
                                        out=out, mask=mask,
-                                       shift_x=shift_x, shift_y=shift_y)
+                                       shift_x=shift_x, shift_y=shift_y,
+                                       out_dtype=np.double)
     else:
         return _apply_scalar_per_pixel_3D(generic_cy._otsu_3D, image,
                                           selem, out=out, mask=mask,
